refactor space container to its own component
