+++ app/packages/core/src/components/MainSpace/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MainSpace";
