+++ app/packages/core/src/components/SamplesContainer.tsx
@@ -1,12 +1,10 @@
-import React, { useCallback, useEffect, useRef } from "react";
+import React, { useCallback } from "react";
 import { useRecoilValue } from "recoil";
 import { Controller } from "@react-spring/web";
 import styled from "styled-components";
 import Sidebar, { Entries } from "./Sidebar";
 import * as fos from "@fiftyone/state";
-import { SpacesRoot, usePanelsState, useSpaces } from "@fiftyone/spaces";
-import { useSessionSpaces } from "@fiftyone/state";
-import { size, isEqual } from "lodash";
+import MainSpace from "./MainSpace";
 
 const Container = styled.div`
   display: flex;
@@ -19,13 +17,6 @@ const Container = styled.div`
 function SamplesContainer() {
   const showSidebar = useRecoilValue(fos.sidebarVisible(false));
   const disabled = useRecoilValue(fos.disabledPaths);
-  const [sessionSpaces, setSessionSpaces, sessionPanelsState] =
-    useSessionSpaces();
-  const { spaces, updateSpaces } = useSpaces("main", sessionSpaces);
-  const [panelsState, setPanelsState] = usePanelsState();
-  const oldSpaces = useRef(sessionSpaces);
-  const oldPanelsState = useRef(panelsState);
-  const isMounted = useRef(false);
 
   const renderGridEntry = useCallback(
     (
@@ -141,41 +132,10 @@ function SamplesContainer() {
     },
     []
   );
-
-  useEffect(() => {
-    if (!spaces.equals(sessionSpaces)) {
-      updateSpaces(sessionSpaces);
-    }
-  }, [sessionSpaces]);
-
-  useEffect(() => {
-    if (size(sessionPanelsState)) {
-      setPanelsState(sessionPanelsState);
-    }
-  }, [sessionPanelsState]);
-
-  useEffect(() => {
-    if (!isMounted.current) {
-      isMounted.current = true;
-      return;
-    }
-    const serializedSpaces = spaces.toJSON();
-    const spacesUpdated =
-      !spaces.equals(sessionSpaces) && !spaces.equals(oldSpaces.current);
-    const panelsStateUpdated =
-      !isEqual(sessionPanelsState, panelsState) &&
-      !isEqual(panelsState, oldPanelsState.current);
-    if (spacesUpdated || panelsStateUpdated) {
-      setSessionSpaces(serializedSpaces, panelsState);
-    }
-    oldSpaces.current = serializedSpaces;
-    oldPanelsState.current = panelsState;
-  }, [spaces, panelsState]);
-
   return (
     <Container>
       {showSidebar && <Sidebar render={renderGridEntry} modal={false} />}{" "}
-      <SpacesRoot id={"main"} />
+      <MainSpace />
     </Container>
   );
 }
